refactoring(integrations): removed renderIntegrationField function, IntegrationFormField component created instead
