+++ src/features/Integrations/index.ts
@@ -1,2 +1,3 @@
 export { type IntegrationsProps, Integrations } from "./Integrations";
 export * from "./integrationsSlice";
+export { IntegrationFormField } from "./IntegrationFormField";
